presidio-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pydantic==2.5.0
PyYAML==6.0.1

